feat: implement sign-in functionality and related types; refactor sign-up logic to include auto-login
