+++ src/features/sign-up/hooks/useSignUp.ts
@@ -2,35 +2,20 @@
 
 import { useMutation } from "@tanstack/react-query";
 import type { AxiosError } from "axios";
-import {
-  signUp,
-  signIn,
-  persistTokens,
-  type SignUpPayload,
-  type AuthTokens,
-} from "../services";
-
-type MutationInput = SignUpPayload;
-type MutationOutput = AuthTokens;
+import { signUp, type SignUpPayload } from "../services";
+import { signIn } from "@/features/sign-in/services/auth";
+import { persistTokens } from "@/shared/helpers/auth";
+import type { AuthTokens } from "@/shared/types/auth";
 
 export function useSignUpAndAutoLogin() {
   return useMutation<AuthTokens, AxiosError, SignUpPayload>({
+    mutationKey: ["auth", "signup"],
+    retry: false,
     mutationFn: async (p) => {
       await signUp(p);
-      try {
-        const t = await signIn({ username: p.username, password: p.password });
-        persistTokens(t);
-        return t;
-      } catch (e: any) {
-        const email = p.company?.contactEmail;
-        const status = e?.response?.status;
-        if (email && (status === 400 || status === 401 || status === 404)) {
-          const t = await signIn({ username: email, password: p.password });
-          persistTokens(t);
-          return t;
-        }
-        throw e;
-      }
+      const t = await signIn({ username: p.username, password: p.password });
+      persistTokens(t);
+      return t;
     },
   });
 }
