+++ src/features/sign-up/services/auth.ts
@@ -10,17 +10,6 @@ export type SignUpPayload = {
   };
 };
 
-export type SignInPayload = {
-  username: string;
-  password: string;
-};
-
-export type AuthTokens = {
-  accessToken?: string;
-  refreshToken?: string;
-  [k: string]: unknown;
-};
-
 export async function signUp(payload: SignUpPayload) {
   const clean: SignUpPayload = {
     username: payload.username.trim(),
@@ -33,19 +22,4 @@ export async function signUp(payload: SignUpPayload) {
   };
   const { data } = await api.post("/auth/signup", clean);
   return data;
-}
-export async function signIn(body: SignInPayload) {
-  const { data } = await api.post("/auth/signin", {
-    username: body.username,
-    password: body.password,
-  });
-  return data;
-}
-
-export function persistTokens(tokens: AuthTokens) {
-  if (typeof window === "undefined") return;
-  if (tokens.accessToken)
-    localStorage.setItem("accessToken", tokens.accessToken);
-  if (tokens.refreshToken)
-    localStorage.setItem("refreshToken", tokens.refreshToken);
 }
