+++ src/features/sign-up/services/index.ts
@@ -1,1 +1,2 @@
 export * from "./auth";
+export * from "./errorMap";
