+++ src/lib/axios.ts
@@ -1,13 +1,14 @@
 import axios from "axios";
+import { getAccessToken } from "@/shared/helpers/auth";
 
 export const api = axios.create({
   baseURL: process.env.NEXT_PUBLIC_API_BASE_URL ?? "http://141.11.156.52:3203",
   headers: { "Content-Type": "application/json" },
+  timeout: 15_000,
 });
 
 api.interceptors.request.use((config) => {
-  const token =
-    typeof window !== "undefined" ? localStorage.getItem("accessToken") : null;
+  const token = getAccessToken();
   if (token) config.headers.Authorization = `Bearer ${token}`;
   return config;
 });
