+++ src/shared/types/auth.ts
@@ -1,0 +1,5 @@
+export type AuthTokens = {
+  accessToken?: string;
+  refreshToken?: string;
+  [k: string]: unknown;
+};
